feat: Apply in RenderSections
